Update: Separate checkboxLabel to its own component
The reason is so that we can use the styled label and ensure it is
styled the same as the other checkbox's labels
