+++ packages/es-components/src/components/controls/checkbox/CheckboxLabel.md
@@ -1,0 +1,7 @@
+The styled label used for the checkbox control
+
+```
+<CheckboxLabel>
+  Hi there
+</CheckboxLabel>
+```
